+++ notebooks/hello_revodata.py
@@ -1,9 +1,0 @@
-# Databricks notebook source
-# MAGIC %md
-# MAGIC # Notebook Task
-
-# COMMAND ----------
-
-print("Hello, RevoData! \nNotebook task is successfully completed.")
-
-# COMMAND ----------