+++ CHANGELOG.md
@@ -8,3 +8,32 @@ All notable changes to this project will be documented in this file.
 
 - Initial commit
 
+## v1.1.0 (2025-09-24)
+
+### Feat
+
+- add new diagnostic notebooks
+- add diagnostic and ingestion modules
+- **hierarchy**: add script to generate Mermaid hierarchy from json
+- **get-databricks-creds**: add shell script to extact Databricks credentials
+- add workspace_information diagnostic notebook
+- add Jinja templates
+
+### Fix
+
+- correct dependabot.yml configuration structure
+- improve main CLI and workspace information
+- dlt_flows
+- **get-databricks-creds**: include CSV generation
+- **generate-hierarchy**: add argument parser options
+- **databricks-credentials**: debug `workspace_information.py`
+- **uv**: remove all remaining poetry references
+- dlt_flows
+- jinja append
+- **ruff**: fixed a bug in Ruff configuration
+
+## v1.0.0 (2024-11-21)
+
+### Feat
+
+- initial commit